Fixed seq, swisstm, tl2, tinystm, norec (rstm), missing RTM

+++ ssca2/findSubGraphs.c
@@ -6,48 +6,48 @@
  *
  * For the license of bayes/sort.h and bayes/sort.c, please see the header
  * of the files.
- * 
+ *
  * ------------------------------------------------------------------------
- * 
+ *
  * For the license of kmeans, please see kmeans/LICENSE.kmeans
- * 
+ *
  * ------------------------------------------------------------------------
- * 
+ *
  * For the license of ssca2, please see ssca2/COPYRIGHT
- * 
+ *
  * ------------------------------------------------------------------------
- * 
+ *
  * For the license of lib/mt19937ar.c and lib/mt19937ar.h, please see the
  * header of the files.
- * 
+ *
  * ------------------------------------------------------------------------
- * 
+ *
  * For the license of lib/rbtree.h and lib/rbtree.c, please see
  * lib/LEGALNOTICE.rbtree and lib/LICENSE.rbtree
- * 
+ *
  * ------------------------------------------------------------------------
- * 
+ *
  * Unless otherwise noted, the following license applies to STAMP files:
- * 
+ *
  * Copyright (c) 2007, Stanford University
  * All rights reserved.
- * 
+ *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are
  * met:
- * 
+ *
  *     * Redistributions of source code must retain the above copyright
  *       notice, this list of conditions and the following disclaimer.
- * 
+ *
  *     * Redistributions in binary form must reproduce the above copyright
  *       notice, this list of conditions and the following disclaimer in
  *       the documentation and/or other materials provided with the
  *       distribution.
- * 
+ *
  *     * Neither the name of Stanford University nor the names of its
  *       contributors may be used to endorse or promote products derived
  *       from this software without specific prior written permission.
- * 
+ *
  * THIS SOFTWARE IS PROVIDED BY STANFORD UNIVERSITY ``AS IS'' AND ANY
  * EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
  * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR
@@ -107,14 +107,14 @@ findSubGraphs0 (void* argPtr)
 
     for (i = i_start; i < i_stop; i++) {
         if (i < maxIntWtListSize) {
-            long j;
+            unsigned long j;
             for (j = 0; j < GPtr->numVertices; j++) {
                 intWtVList[i*(GPtr->numVertices)+j].num = 0;
                 intWtVList[i*(GPtr->numVertices)+j].depth = 0;
             }
         } else {
             long t = i - maxIntWtListSize;
-            long j;
+            unsigned long j;
             for (j = 0; j < GPtr->numVertices; j++) {
                 strWtVList[t*(GPtr->numVertices)+j].num = 0;
                 strWtVList[t*(GPtr->numVertices)+j].depth = 0;
@@ -134,7 +134,7 @@ findSubGraphs0 (void* argPtr)
 
     for (i = i_start; i < i_stop; i++) {
 
-        long k;
+        unsigned long k;
         for (k = 0; k < GPtr->numVertices; k++) {
             visited[k] = 0;
         }
@@ -155,7 +155,7 @@ findSubGraphs0 (void* argPtr)
             long currIndex = 1;
 
             while ((depth < SUBGR_EDGE_LENGTH) ||
-                   (verticesVisited == GPtr->numVertices))
+                   (verticesVisited == (long)GPtr->numVertices))
             {
                 long intWtListIndex = i*(GPtr->numVertices)+currIndex;
                 depth = intWtVList[intWtListIndex].depth + 1;
@@ -173,7 +173,7 @@ findSubGraphs0 (void* argPtr)
                     }
                 }
                 if ((currIndex < verticesVisited - 1) &&
-                    (verticesVisited < GPtr->numVertices))
+                    (verticesVisited < (long)GPtr->numVertices))
                 {
                     currIndex++;
                     depth = intWtVList[i*(GPtr->numVertices)+currIndex].depth;
@@ -202,7 +202,7 @@ findSubGraphs0 (void* argPtr)
             long currIndex = 1;
 
             while ((depth < SUBGR_EDGE_LENGTH) ||
-                   (verticesVisited == GPtr->numVertices))
+                   (verticesVisited == (long)GPtr->numVertices))
             {
                 long strWtVListIndex = t*(GPtr->numVertices)+currIndex;
                 depth = strWtVList[strWtVListIndex].depth + 1;
@@ -268,7 +268,7 @@ findSubGraphs1 (void* argPtr)
 
     for (i = i_start; i < i_stop; i++) {
 
-        long k;
+        unsigned long k;
         for (k = 0; k < GPtr->numVertices; k++)  {
             visited[k] = 'u';
         }
@@ -295,7 +295,7 @@ findSubGraphs1 (void* argPtr)
             long currIndex = 1;
 
             while ((startV->depth < SUBGR_EDGE_LENGTH) ||
-                   (verticesVisited == GPtr->numVertices))
+                   (verticesVisited == (long)GPtr->numVertices))
             {
                 depth = startV->depth + 1;
                 long j;
@@ -313,12 +313,13 @@ findSubGraphs1 (void* argPtr)
                     }
                 }
 
-                if ((currIndex < verticesVisited - 1) && (verticesVisited < GPtr->numVertices)) {
-                currIndex++;
-                startV = startV->next;
-
+                if ((currIndex < verticesVisited - 1) &&
+                    (verticesVisited < (long)GPtr->numVertices))
+                {
+                   currIndex++;
+                   startV = startV->next;
                 } else {
-                break;
+                  break;
                 }
             }
             currV->next = NULL;
@@ -346,7 +347,7 @@ findSubGraphs1 (void* argPtr)
             long currIndex = 1;
 
             while ((startV->depth < SUBGR_EDGE_LENGTH) ||
-                   (verticesVisited == GPtr->numVertices))
+                   (verticesVisited == (long)GPtr->numVertices))
             {
                 depth = startV->depth + 1;
 
@@ -366,7 +367,7 @@ findSubGraphs1 (void* argPtr)
                 }
 
                 if ((currIndex < verticesVisited - 1) &&
-                    (verticesVisited < GPtr->numVertices))
+                    (verticesVisited < (long)GPtr->numVertices))
                 {
                     currIndex++;
                     startV = startV->next;
@@ -464,7 +465,7 @@ findSubGraphs2 (void* argPtr)
 
     for (i = i_start; i < i_stop; i++) {
 
-        long k;
+        unsigned long k;
         for (k = 0; k < GPtr->numVertices; k++) {
             visited[k] = 'u';
         }
@@ -487,7 +488,7 @@ findSubGraphs2 (void* argPtr)
             long currIndex = 1;
 
             while ((depth < SUBGR_EDGE_LENGTH) ||
-                   (verticesVisited == GPtr->numVertices))
+                   (verticesVisited == (long)GPtr->numVertices))
             {
                 long currVListX = currIndex / arraySize;
                 long currVListY = currIndex % arraySize;
@@ -514,7 +515,7 @@ findSubGraphs2 (void* argPtr)
                 if (((float) verticesVisited / (float) arraySize) > 0.5) {
                     /* create a new sub-array */
                     if (intWtVDList[i].numArrays !=
-                        (verticesVisited/arraySize + 2))
+                        (unsigned long)(verticesVisited/arraySize + 2))
                     {
                         intWtVDList[i].numArrays++;
                         intWtVDList[i].vList[intWtVDList[i].numArrays-1] =
@@ -525,7 +526,7 @@ findSubGraphs2 (void* argPtr)
                 }
 
                 if ((currIndex < verticesVisited - 1) &&
-                    (verticesVisited < GPtr->numVertices))
+                    (verticesVisited < (long)GPtr->numVertices))
                 {
                     currIndex++;
                     long vListX = currIndex / arraySize;
@@ -556,7 +557,7 @@ findSubGraphs2 (void* argPtr)
             long currIndex = 1;
 
             while ((depth < SUBGR_EDGE_LENGTH) ||
-                   (verticesVisited == GPtr->numVertices))
+                   (verticesVisited == (long)GPtr->numVertices))
             {
                 long currVListX = currIndex / arraySize;
                 long currVListY = currIndex % arraySize;
@@ -583,7 +584,7 @@ findSubGraphs2 (void* argPtr)
                 if (((float)verticesVisited /(float) arraySize) > 0.5) {
                     /* create a new sub-array */
                     if (strWtVDList[t].numArrays !=
-                        (verticesVisited/arraySize + 2))
+                        (unsigned long)(verticesVisited/arraySize + 2))
                     {
                         strWtVDList[t].numArrays++;
                         strWtVDList[t].vList[strWtVDList[t].numArrays-1] =
@@ -594,7 +595,7 @@ findSubGraphs2 (void* argPtr)
                 }
 
                 if ((currIndex < verticesVisited - 1) &&
-                    (verticesVisited < GPtr->numVertices))
+                    ((unsigned long)verticesVisited < GPtr->numVertices))
                 {
                     currIndex++;
                     long currVListX = currIndex / arraySize;
